+++ src/Farmer/Builders/Builders.Vm.fs
@@ -41,8 +41,8 @@ type VmConfig =
       PublicIp: ResourceRef<VmConfig> option
       IpAllocation: PublicIpAddress.AllocationMethod option
       PrivateIpAllocation: PrivateIpAddress.AllocationMethod option
-
       Identity : Identity.ManagedIdentity
+      NetworkSecurityGroup: LinkedResource option
 
       Tags: Map<string,string> }
 
@@ -87,6 +87,7 @@ type VmConfig =
 
             let vnetName = this.VNet.resourceId(this).Name
             let subnetName = this.Subnet.resourceId this
+            let nsgId = this.NetworkSecurityGroup |> Option.map(fun nsg -> nsg.ResourceId)
 
             // NIC
             { Name = this.NicName.Name
@@ -98,6 +99,7 @@ type VmConfig =
                         |> Option.map (fun x -> x.toLinkedResource this) |} ]
               VirtualNetwork = vnetName
               PrivateIpAllocation = this.PrivateIpAllocation
+              NetworkSecurityGroup = nsgId
               Tags = this.Tags }
 
             // VNET
@@ -109,7 +111,7 @@ type VmConfig =
                   Subnets = [
                       { Name = subnetName.Name
                         Prefix = this.SubnetPrefix
-                        NetworkSecurityGroup = None
+                        NetworkSecurityGroup = nsgId |> Option.map(fun x -> Managed x)
                         Delegations = []
                         ServiceEndpoints = []
                         AssociatedServiceEndpointPolicies = []
@@ -192,6 +194,7 @@ type VirtualMachineBuilder() =
           PublicIp = automaticPublicIp
           IpAllocation = None
           PrivateIpAllocation = None
+          NetworkSecurityGroup = None
           Tags = Map.empty }
 
     member __.Run (state:VmConfig) =
@@ -301,6 +304,22 @@ type VirtualMachineBuilder() =
     member _.PrivateIpAllocation(state: VmConfig, ref: PrivateIpAddress.AllocationMethod Option) = { state with PrivateIpAllocation = ref}
     member _.PrivateIpAllocation(state: VmConfig, ref: PrivateIpAddress.AllocationMethod) = { state with PrivateIpAllocation = Some ref}
 
+    /// Sets the network security group
+    [<CustomOperation "network_security_group">]
+    member _.NetworkSecurityGroup(state:VmConfig, nsg:IArmResource) =
+        { state with NetworkSecurityGroup = Some (Managed nsg.ResourceId) }
+    member _.NetworkSecurityGroup(state:VmConfig, nsg:ResourceId) =
+        { state with NetworkSecurityGroup = Some (Managed nsg) }
+    member _.NetworkSecurityGroup(state:VmConfig, nsg:NsgConfig) =
+        { state with NetworkSecurityGroup = Some (Managed (nsg :> IBuilder).ResourceId) }
+    /// Links the VM to an existing network security group.
+    [<CustomOperation "link_to_network_security_group">]
+    member _.LinkToNetworkSecurityGroup(state:VmConfig, nsg:IArmResource) =
+        { state with NetworkSecurityGroup = Some (Unmanaged (nsg.ResourceId)) }
+    member _.LinkToNetworkSecurityGroup(state:VmConfig, nsg:ResourceId) =
+        { state with NetworkSecurityGroup = Some (Unmanaged nsg) }
+    member _.LinkToNetworkSecurityGroup(state:VmConfig, nsg:NsgConfig) =
+        { state with NetworkSecurityGroup = Some (Unmanaged (nsg :> IBuilder).ResourceId) }
 
 
 let vm = VirtualMachineBuilder()